filesync, fix wopi ttl

+++ modules/filesync/config/tablemodel.php
@@ -101,7 +101,7 @@ $tbs[] = $tb_filesync_store_file_rev;
 $tb_wopi_token = new TableModel('filesync', 'wopi_access');
 $tb_wopi_token->addColumn('wopi_access_id',   'int', ['key' => 'PRIMARY KEY', 'auto_increment' => true]);
 $tb_wopi_token->addColumn('access_token',     'varchar(1024)');
-$tb_wopi_token->addColumn('access_token_ttl', 'int');
+$tb_wopi_token->addColumn('access_token_ttl', 'bigint');
 $tb_wopi_token->addColumn('base_path',        'varchar(1024)');
 $tb_wopi_token->addColumn('path',             'varchar(1024)');
 $tb_wopi_token->addColumn('user_id',          'int');
